Moved plugins from src to plugins
