+++ RecoMuon/MuonIdentification/plugins/ME0MuonConverter.cc
@@ -3,7 +3,7 @@
  * \author David Nash
  */
 
-#include <RecoMuon/MuonIdentification/interface/ME0MuonConverter.h>
+#include <RecoMuon/MuonIdentification/plugins/ME0MuonConverter.h>
 
 #include <FWCore/PluginManager/interface/ModuleDef.h>
 #include <FWCore/Framework/interface/MakerMacros.h>
